pyproject.toml
[project]
name = "bont"
version = "1.1.0"
description = "A Python module for converting TrueType fonts to bitmap font texture atlas."
authors = [{name="Andrew Kennedy"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fonttools",
    "pillow",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["pygame", "game development"]

[project.urls]
Homepage = "https://github.com/kennedy0/bont"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
